+++ tests/compilers/imports.rs
@@ -341,7 +341,7 @@ fn dynamic_function_with_env_wasmer_env_init_works(config: crate::Config) -> Res
             },
         },
     )?;
-    let f: NativeFunc<(), ()> = instance.exports.get_native_function("main")?;
+    let f: TypedFunction<(), ()> = instance.exports.get_native_function("main")?;
     f.call()?;
     Ok(())
 }
@@ -381,12 +381,12 @@ fn multi_use_host_fn_manages_memory_correctly(config: crate::Config) -> Result<(
     let instance1 = Instance::new(&module, &imports)?;
     let instance2 = Instance::new(&module, &imports)?;
     {
-        let f1: NativeFunc<(), ()> = instance1.exports.get_native_function("main")?;
+        let f1: TypedFunction<(), ()> = instance1.exports.get_native_function("main")?;
         f1.call()?;
     }
     drop(instance1);
     {
-        let f2: NativeFunc<(), ()> = instance2.exports.get_native_function("main")?;
+        let f2: TypedFunction<(), ()> = instance2.exports.get_native_function("main")?;
         f2.call()?;
     }
     drop(instance2);
@@ -425,8 +425,8 @@ fn instance_local_memory_lifetime(config: crate::Config) -> Result<()> {
         },
     };
     let instance = Instance::new(&module, &imports)?;
-    let set_at: NativeFunc<(i32, i32), ()> = instance.exports.get_native_function("set_at")?;
-    let get_at: NativeFunc<i32, i32> = instance.exports.get_native_function("get_at")?;
+    let set_at: TypedFunction<(i32, i32), ()> = instance.exports.get_native_function("set_at")?;
+    let get_at: TypedFunction<i32, i32> = instance.exports.get_native_function("get_at")?;
     set_at.call(200, 123)?;
     assert_eq!(get_at.call(200)?, 123);